+++ git-remote-updater.php
@@ -14,7 +14,7 @@
  * Description:       Allows you to easily update Git Updater repositories in bulk via REST API endpoint updating. Requires Git Updater PRO.
  * Author:            Andy Fragen
  * Author URI:        https://github.com/afragen
- * Version:           2.2.3
+ * Version:           2.2.3.1
  * License:           MIT
  * Network:           true
  * Domain Path:       /languages